:heavy_plus_sign: Enable chunker on RPi boot.

+++ kernel/Heap.h
@@ -3,14 +3,16 @@
 #ifndef __HEAP_H__
 #define __HEAP_H__
 
+#include <cstddef>
+
 namespace Kernel
 {
 	namespace Heap
 	{
 		void Init(void);
 
-		void* Alloc(unsigned long bytes, bool phys = true);
-		void* Alloc(unsigned long bytes, unsigned long align, bool phys = true);
+		void* Alloc(size_t bytes, bool phys = true);
+		void* Alloc(size_t bytes, size_t align, bool phys = true);
 		void Free(void* addr, bool phys = true);
 
 		void ShowMap(void);
